backend/pyproject.toml
[project]
name = "vibe-gtm-backend"
version = "0.1.0"
description = "Vibe GTM backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "firecrawl-py>=4.14.0",
    "openai>=1.0.0",
    "uvicorn[standard]>=0.27.0",
    "motor>=3.3.0",
    "pydantic-settings>=2.0.0",
    "reductoai>=0.15.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]